refactor(jobs): flatten tasks/ into jobs/
Move CleanupOrphanedMedia and SyncLectureMetadata from src/jobs/tasks/
directly into src/jobs/ and delete the tasks/ subdirectory. Update
jobs/index.ts and the two integration tests that import them.

Refs #442
