+++ src/jobs/SyncLectureMetadata.ts
