+++ src/jobs/index.ts
@@ -1,5 +1,5 @@
-import { CleanupOrphanedMedia } from './tasks/CleanupOrphanedMedia'
-import { SyncLectureMetadata } from './tasks/SyncLectureMetadata'
+import { CleanupOrphanedMedia } from './CleanupOrphanedMedia'
+import { SyncLectureMetadata } from './SyncLectureMetadata'
 
 // Export all tasks as an array
 // Note: TrackUsage and ResetUsage tasks are auto-registered by the usagePlugin
