+++ tests/int/sync-lecture-metadata.int.spec.ts
@@ -3,9 +3,8 @@ import type { Payload } from 'payload'
 import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
 
 import type { LectureMetadata } from '@/hooks/lectureHooks'
-import { SyncLectureMetadata } from '@/jobs/tasks/SyncLectureMetadata'
+import { SyncLectureMetadata } from '@/jobs/SyncLectureMetadata'
 import type { Lecture } from '@/payload-types'
-
 
 import { testData } from '../utils/testData'
 import { createTestEnvironment } from '../utils/testHelpers'
